refactor : Changed access modifiers on BaseballGame docs : updated README.md
